refactor: - extract StatusEditor component
