+++ src/ui/StatusEditor.svelte
@@ -1,0 +1,48 @@
+<script lang="ts">
+    import type { Status } from '../Statuses/Status';
+    import type { Task } from '../Task/Task';
+    import type { EditableTask } from './EditableTask';
+    import { labelContentWithAccessKey } from './EditTaskHelpers';
+
+    export let accesskey: string | null;
+    export let editableTask: EditableTask;
+    export let statusOptions: Status[];
+    export let task: Task;
+
+    let statusSymbol = task.status.symbol;
+
+    const _onStatusChange = () => {
+        // Use statusSymbol to find the status to save to editableTask.status
+        const selectedStatus: Status | undefined = statusOptions.find((s) => s.symbol === statusSymbol);
+        if (selectedStatus) {
+            editableTask.status = selectedStatus;
+        } else {
+            console.log(`Error in EditTask: cannot find status with symbol ${statusSymbol}`);
+            return;
+        }
+
+        // Obtain a temporary task with the new status applied, to see what would
+        // happen to the done date:
+        const taskWithEditedStatusApplied = task.handleNewStatus(selectedStatus).pop();
+
+        if (taskWithEditedStatusApplied) {
+            // Update the doneDate field, in case changing the status changed the value:
+            editableTask.doneDate = taskWithEditedStatusApplied.done.formatAsDate();
+            editableTask.cancelledDate = taskWithEditedStatusApplied.cancelled.formatAsDate();
+        }
+    };
+</script>
+
+<label for="status">{@html labelContentWithAccessKey('Status', accesskey)}</label>
+<!-- svelte-ignore a11y-accesskey -->
+<select
+    bind:value={statusSymbol}
+    on:change={_onStatusChange}
+    id="status-type"
+    class="tasks-modal-status-selector"
+    {accesskey}
+>
+    {#each statusOptions as status}
+        <option value={status.symbol}>{status.name} [{status.symbol}]</option>
+    {/each}
+</select>
